requirements.txt
python-telegram-bot==20.8
openai==1.51.2
elevenlabs==0.2.26
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0
audioop-lts==0.2.1
pytz==2025.2 